+++ modules/auxiliary/admin/sap/cve_2020_6207_solman_rce.rb
@@ -3,12 +3,10 @@
 # Current source: https://github.com/rapid7/metasploit-framework
 ##
 
-require 'metasploit/framework/sap_solman/client'
-
 class MetasploitModule < Msf::Auxiliary
 
   include Msf::Exploit::Remote::HttpClient
-  include Metasploit::Framework::SapSolutionManager::Client
+  include Msf::Exploit::Remote::HTTP::SapSolManEemMissAuth
 
   @agents = Array.new # Array of connected agents
