+++ modules/exploits/multi/sap/cve_2020_6207_solman_rs.rb
@@ -3,13 +3,11 @@
 # Current source: https://github.com/rapid7/metasploit-framework
 ##
 
-require 'metasploit/framework/sap_solman/client'
-
 class MetasploitModule < Msf::Exploit::Remote
   Rank = NormalRanking
 
   include Msf::Exploit::Remote::HttpClient
-  include Metasploit::Framework::SapSolutionManager::Client
+  include Msf::Exploit::Remote::HTTP::SapSolManEemMissAuth
 
   @agents = Array.new # Array of connected agents
